requirements.txt
pandas
streamlit
google-generativeai
openpyxl